pagination on setup timers
