+++ src/components/Pomodoro/PomodoroPresets.vue
@@ -1,37 +1,21 @@
 <template>
-  <div class="pomo-presets">
-    <div v-for="t in timers"
-      :class="`preset-box ${(timerSelected === t.id) ? 'bg-primary' : 'bg-background'}`" v-ripple
-      @mouseleave="deleteTimer(undefined)" @click="setTimerPreset(t)">
-      <v-icon v-if="deletable"
-        size="x-small" icon="mdi-delete"
-        @click.stop="deleteTimer(t.id)" class="btn-card"
-        :color="deletingTimer === t.id ? 'red' : undefined" />
-      <v-icon v-else
-        size="x-small" icon="mdi-cog"
-        @click.stop="emit('open-settings-tab', 'pomodoro')" class="btn-card"/>
-
-      <div class="timer-card front">
-        <p class="timer-title">{{ t.title }} </p>
-      </div>
-      <div class="timer-card back" v-if="!t.freeMode">
-        <p class="timer-info">{{ t.studyLength }} {{ $t('pause.timer.studyMinShort') }}</p>
-        <p class="timer-info">{{ t.breakLength }} {{ $t('pause.timer.breakMinShort') }}</p>
-        <p class="timer-info">{{ t.repetitions }} {{ $t('pause.timer.repetitions') }}</p>
-      </div>
-      <div class="timer-card back" v-else>
-        <p class="timer-info timer-info-freemode">{{ $t('pause.timer.freeMode') }}</p>
-      </div>
-    </div>
+  <div :class="`pomo-presets ${top3 ? 'top3' : ''}`">
+    <div class="arrow" v-ripple @click="decreasePage()" v-if="top3"
+    ><v-icon icon="mdi-chevron-left" :disabled="page === 0"/></div>
+    <PomodoroPresetsTile
+      :timers="timers"
+      :deletable="deletable"
+      @open-settings-tab="emit('open-settings-tab', $event)" />
+    <div class="arrow" v-ripple @click="increasePage()" v-if="top3"
+    ><v-icon icon="mdi-chevron-right" :disabled="timerStore.timers?.length! <= (page + 1) * 3" /></div>
   </div>
 </template>
 
 <script lang="ts" setup>
 import { computed, ref } from 'vue'
 import { useTimerStore } from "@/stores/settings/timer";
-import { useSettingsStore } from "@/stores/settings";
-import type { Timer } from '@/types';
-const settingsStore = useSettingsStore();
+import PomodoroPresetsTile from './PomodoroPresetsTile.vue';
+
 const timerStore = useTimerStore();
 const emit = defineEmits<{
   (e: 'open-settings-tab', value: string): void
@@ -40,160 +24,47 @@ const props = defineProps<{
   deletable?: boolean,
   top3?: boolean
 }>()
+const timers = computed(() => props.top3 ? timerStore.timers.slice(page.value * 3, (page.value + 1) * 3) : timerStore.timers);
+const page = ref(0);
 
-const timers = computed(() => props.top3 ? timerStore.timers.slice(0, 3) : timerStore.timers);
-
-const timerSelected = computed(() => timerStore.selectedTimerId ??
-  timerStore.timers.find(timer => {
-    const numberOfBreak = timer.repetitions - 1;
-    const breaksLength = timer.breakLength * numberOfBreak;
-    const totalLength = (timer.studyLength * timer.repetitions) + breaksLength;
-
-    return (
-      settingsStore.settings!.pomodoro!.totalLength === totalLength
-      && settingsStore.settings!.pomodoro!.breaksLength === breaksLength
-      && settingsStore.settings!.pomodoro!.numberOfBreak === numberOfBreak
-    );
-  })?.id
-);
-
-const selectedTimer = ref<Timer>({
-  title: '',
-  studyLength: settingsStore.settings!.pomodoro!.totalLength - settingsStore.settings!.pomodoro!.breaksLength,
-  breakLength: settingsStore.settings!.pomodoro!.breaksLength / settingsStore.settings!.pomodoro!.numberOfBreak,
-  repetitions: settingsStore.settings!.pomodoro!.numberOfBreak + 1,
-  freeMode: settingsStore.settings!.pomodoro!.freeMode
-});
-function setTimerPreset(timer: Timer) {
-  selectedTimer.value.breakLength = timer.breakLength;
-  selectedTimer.value.studyLength = timer.studyLength;
-  selectedTimer.value.repetitions = timer.repetitions;
-  selectedTimer.value.freeMode = timer.freeMode;
-  timerStore.selectedTimerId = timer.id ?? null;
-  setTimer();
-}
-function setTimer() {
-  const numberOfBreak = selectedTimer.value.repetitions - 1;
-  const breaksLength = selectedTimer.value.breakLength * numberOfBreak;
-  const totalLength = (selectedTimer.value.studyLength * selectedTimer.value.repetitions) + breaksLength;
-
-  settingsStore.settings!.pomodoro!.totalLength = totalLength;
-  settingsStore.settings!.pomodoro!.breaksLength = breaksLength;
-  settingsStore.settings!.pomodoro!.numberOfBreak = numberOfBreak;
-  settingsStore.settings!.pomodoro!.freeMode = selectedTimer.value.freeMode;
-}
-
-const deletingTimer = ref<number | null>(null);
-function deleteTimer(id: number | undefined) {
-  if (id === undefined) {
-    deletingTimer.value = null;
-    return;
+function increasePage() {
+  if (timerStore.timers?.length! > (page.value + 1) * 3) {
+    page.value++;
   }
-
-  if (deletingTimer.value === id) {
-    deletingTimer.value = null;
-    timerStore.deleteTimer(id);
-  } else {
-    deletingTimer.value = id;
+}
+function decreasePage() {
+  if (page.value > 0) {
+    page.value--;
   }
 }
 </script>
 <style lang="scss" scoped>
 .pomo-presets {
   display: grid;
-  grid-template-columns: repeat(3, 1fr);
+  grid-template-columns: 1fr 1fr 1fr;
   gap: 1rem;
-  place-items: center;
   margin: 1.5rem 0;
   cursor: pointer;
   overflow-x: auto;
   overflow-y: hidden;
 
-  .preset-box {
-    border-radius: 1rem;
-    width: 100%;
-    height: 5rem;
-    position: relative !important;
-
-    .btn-card {
-      display: none;
-      position: absolute;
-      bottom: 5px;
-      right: 5px;
-      opacity: 0.7;
-    }
-
-    .timer-card {
+  &.top3 {
+    grid-template-columns: 1.5rem 1fr 1fr 1fr 1.5rem;
+  }
+  .arrow {
       display: flex;
-      flex-direction: column;
-      padding: 1rem;
-      width: 100%;
       justify-content: center;
       align-items: center;
-      height: 100%;
-    }
+      cursor: pointer;
 
-    .front {
-      display: flex;
-    }
-
-    .back {
-      display: none;
-
-    }
-
-    @media (hover: hover) {
-      &:hover {
-        border: 2px solid rgb(var(--v-theme-primary));
-
-        .front {
-          display: none;
-        }
-
-        .back {
-          display: flex;
-        }
-
-        .btn-card {
-          display: block;
-
-          &:hover {
-            opacity: 1;
-          }
-        }
-      }
-    }
-
-    p {
-      text-align: center;
-      width: 100%;
-
-      &.timer-title {
-        font-size: 1.1rem;
-        line-height: 1.1rem;
-        font-weight: 500;
+      &:first-child {
+        border-radius: 0.5rem;
       }
 
-      &.timer-info {
-        font-size: 0.8rem;
-        text-wrap: nowrap;
-        &.timer-info-freemode {
-          text-wrap: wrap;
-        }
+      &:last-child {
+        border-radius: 0.5rem;
+        grid-column-start: 5;
       }
     }
-  }
-
-  .add-box {
-    grid-column-start: 1;
-    grid-column-end: 4;
-    padding: 0;
-
-    .add-btn {
-      width: 100%;
-      height: 3rem;
-      border-radius: 1rem;
-    }
-  }
 }
 </style>
